+++ dataSources/parquet/src/main/java/cn/edu/tsinghua/iginx/parquet/io/parquet/ParquetIndex.java
@@ -17,8 +17,8 @@
 package cn.edu.tsinghua.iginx.parquet.io.parquet;
 
 import cn.edu.tsinghua.iginx.engine.shared.operator.filter.Filter;
-import cn.edu.tsinghua.iginx.parquet.db.common.scanner.IteratorScanner;
 import cn.edu.tsinghua.iginx.parquet.db.lsm.api.Scanner;
+import cn.edu.tsinghua.iginx.parquet.db.lsm.scanner.IteratorScanner;
 import cn.edu.tsinghua.iginx.parquet.io.FileIndex;
 import cn.edu.tsinghua.iginx.parquet.io.FileMeta;
 import java.io.IOException;
